restructured app to remove context, new location data is now recieved by map component and map changes location
