+++ src/components/Map.js
@@ -1,20 +1,29 @@
-import React, {useContext} from 'react';
+import React from 'react';
 import GoogleMapReact from 'google-map-react';
 import MyLocation from './MyLocation';
-import LocationContext  from './LocationContext'
 
-function Map(){
+
+function Map(props){
+    const {data} = props
+    const defaultProps = {
+        center: {
+          lat: 59.95,
+          lng: 30.33
+        },
+        zoom: 11
+      };
     
-    const contextData=useContext(LocationContext)
-    console.log('location context:', contextData)
     return (
         <div style={{ height: '100vh', width: '100%' }}>
           <GoogleMapReact
             bootstrapURLKeys={{ key:'AIzaSyA_jF-TPUl8qTMZ3BKFTrFOolH9wR7NOz4'}}
-            defaultCenter={contextData.center}
-            defaultZoom={contextData.zoom}
+            defaultCenter={defaultProps.center}
+            defaultZoom={defaultProps.zoom}
             >
-            <MyLocation/>
+            <MyLocation
+            lat={data.center.lat}
+            lng={data.center.lng}
+            />
             </GoogleMapReact> 
             
         </div>
